+++ src/valueObject/UserId.ts
@@ -1,5 +1,5 @@
 import { PropertyValue } from "@notionhq/client/build/src/api-types";
-import { User } from "../entity/User";
+import { User } from "../entity";
 import { PrimitiveValueObject } from "./PrimitiveValueObject";
 
 export class UserId extends PrimitiveValueObject<string> {
